# apps/api/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Social Media SaaS API
  version: 1.0.0
  description: |
    Multi-tenant social media automation platform.

    ## Quick Start
    1. Call **POST /test/seed** → returns a brand-new workspace + owner token (no setup needed).
    2. Click the **Authorize** button at the top and paste the `token` from the response.
    3. All other endpoints (with the lock icon) are now callable directly via **Try it out**.

    Every example body below contains dummy data that satisfies validation; just press Send.

    ## Response envelope
    Success: `{ "success": true, "data": ..., "requestId": "uuid" }`<br/>
    Error:   `{ "success": false, "code": "ERROR_CODE", "message": "...", "requestId": "uuid" }`
servers:
  - url: http://localhost:4000/api/v1
    description: Local dev
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        success: { type: boolean, example: false }
        code:    { type: string, example: VALIDATION_ERROR }
        message: { type: string }
        requestId: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Authenticated but role not allowed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource does not exist (or belongs to a different tenant)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    ValidationError:
      description: Body failed Zod validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Bootstrap a new workspace + first owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, lastName, email, password, workspace]
              properties:
                firstName: { type: string, minLength: 2, maxLength: 60 }
                lastName:  { type: string, minLength: 2, maxLength: 60 }
                email:     { type: string, format: email }
                password:
                  type: string
                  minLength: 12
                  description: ≥12 chars; must include lowercase, uppercase, digit, symbol
                workspace:
                  type: object
                  required: [name, slug]
                  properties:
                    name: { type: string, minLength: 3, maxLength: 80 }
                    slug:
                      type: string
                      pattern: '^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$'
            example:
              firstName: Acme
              lastName: Owner
              email: owner-demo@acme.test
              password: StrongP@ssw0rd!
              workspace:
                name: Acme Inc
                slug: acme-demo-001
      responses:
        '201': { description: Created }
        '409': { description: Slug or email already in use }
        '400': { $ref: '#/components/responses/ValidationError' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Login → JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string }
            example:
              email: owner-demo@acme.test
              password: StrongP@ssw0rd!
      responses:
        '200': { description: Ok }
        '401': { description: Invalid credentials }
  /auth/me:
    get:
      tags: [Auth]
      summary: Current authenticated user
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Ok }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ─────────────────────────────── USERS ─────────────────────────────
  /users:
    get:
      tags: [Users]
      summary: List users in caller's workspace
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Ok }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Users]
      summary: Invite a teammate (owner/admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, lastName, email, password]
              properties:
                firstName: { type: string }
                lastName:  { type: string }
                email:     { type: string, format: email }
                password:  { type: string, minLength: 12 }
                role:      { type: string, enum: [admin, editor, viewer], default: viewer }
            example:
              firstName: Eve
              lastName: Editor
              email: eve-demo@acme.test
              password: AnotherP@ss123!
              role: editor
      responses:
        '201': { description: Created }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: Email already in use }
  /users/{id}:
    get:
      tags: [Users]
      summary: Get a user by id (within caller's workspace)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: Ok }
        '404': { $ref: '#/components/responses/NotFound' }

  # ────────────────────────── WORKSPACES ─────────────────────────────
...